# packages/api-contract/openapi.yaml
openapi: 3.1.0
...
info:
  title: ModuleOS API
  version: 0.1.0
  description: >-
    The versioned HTTP contract for the ModuleOS Genesis control plane. Runtime
    state converges asynchronously; readiness reports dependency availability.
  contact:
    name: ModuleOS project
    url: https://github.com/mevlutkural/moduleos
  license:
    name: MIT
    identifier: MIT
servers:
  - url: /api/v1
    description: ModuleOS control-plane API
...
security:
  - bearerAuth: []
paths:
...
  /apps:
    parameters:
      - $ref: "#/components/parameters/RequestID"
    post:
      operationId: createApplication
      tags: [Applications]
      summary: Create an application
      description: Persists application desired state and queues asynchronous reconciliation.
      requestBody:
        description: Image-based application desired state.
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/CreateApplicationRequest"}
            example: {name: api, project_slug: root, image: "nginx:1.27", replicas: 2}
      responses:
        "201":
          description: Application intent was created.
          headers:
            X-Request-ID: {$ref: "#/components/headers/RequestID"}
            Cache-Control: {$ref: "#/components/headers/NoStore"}
            Location: {$ref: "#/components/headers/Location"}
            ETag: {$ref: "#/components/headers/ApplicationETag"}
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Application"}
              examples:
                current: {$ref: "#/components/examples/Application"}
        "400": {$ref: "#/components/responses/BadRequest"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "409": {$ref: "#/components/responses/Conflict"}
        "413": {$ref: "#/components/responses/PayloadTooLarge"}
        "415": {$ref: "#/components/responses/UnsupportedMediaType"}
        "422": {$ref: "#/components/responses/ValidationError"}
        "500": {$ref: "#/components/responses/InternalError"}
        "503": {$ref: "#/components/responses/ServiceUnavailable"}
    get:
      operationId: listApplications
      tags: [Applications]
      summary: List applications
      responses:
        "200":
          description: Bounded application collection.
          headers:
            X-Request-ID: {$ref: "#/components/headers/RequestID"}
            Cache-Control: {$ref: "#/components/headers/NoStore"}
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ApplicationCollection"}
              example: {applications: [], total: 0}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "500": {$ref: "#/components/responses/InternalError"}
        "503": {$ref: "#/components/responses/ServiceUnavailable"}
...
components:
  examples:
    Application:
      summary: Redacted application state
      description: Example application representation with secret values and host paths removed.
      value:
        id: 20000000-0000-4000-8000-000000000001
        project_id: 00000000-0000-0000-0000-000000000001
        name: api
        source_type: image
        image: "nginx:1.27"
        status: created
        replicas: 2
        desired_run_state: running
        desired_generation: 1
        observed_generation: 0
        observed_state: pending
        reconcile_retryable: false
        env_vars: {SECRET: {redacted: true}}
        ports: [{container_port: 8080, published_port: 0, protocol: tcp, publish_mode: ingress}]
        volumes: [{target: /data, read_only: true, source_redacted: true}]
        expose: false
        ingress_container_port: 0
        created_at: "2026-10-09T09:00:00Z"
        updated_at: "2026-10-09T09:00:00Z"
  parameters:
    RequestID:
      name: X-Request-ID
      in: header
      required: false
      description: >-
        Optional canonical UUID supplied by the caller for end-to-end request
        correlation. Invalid values are replaced by the control plane.
      schema:
        type: string
        format: uuid
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >-
        ModuleOS administrator API key. System routes explicitly opt out. A
        daemon explicitly validated as development may run protected operations
        without authentication only when no API key is configured; production
        always requires a key of at least 32 bytes.
  headers:
    RequestID:
      description: Canonical UUID used to correlate the request with structured logs.
      required: true
      schema:
        type: string
        format: uuid
    NoStore:
      description: Prevents caching of process and dependency status.
      required: true
      schema:
        type: string
        const: no-store
...
    Location:
      description: Canonical relative API path for the created resource.
      required: true
      schema:
        type: string
        pattern: "^/api/v1/"
    ApplicationETag:
      description: Strong ETag containing the current positive desired generation.
      required: true
      schema:
        type: string
        pattern: '^"[1-9][0-9]*"$'
  responses:
    BadRequest:
      $ref: "#/components/responses/GenericBadRequest"
    GenericBadRequest:
      description: Request syntax, JSON, media parameters, or a path value is malformed.
      headers:
        X-Request-ID: {$ref: "#/components/headers/RequestID"}
        Cache-Control: {$ref: "#/components/headers/NoStore"}
      content:
        application/json:
          schema: {$ref: "#/components/schemas/Error"}
    Unauthorized:
      description: A required bearer credential is missing or invalid.
      headers:
        X-Request-ID: {$ref: "#/components/headers/RequestID"}
        Cache-Control: {$ref: "#/components/headers/NoStore"}
        WWW-Authenticate:
          description: Authentication scheme required for protected operations.
          required: true
          schema: {type: string, const: Bearer}
      content:
        application/json:
          schema: {$ref: "#/components/schemas/Error"}
...
    Conflict:
      description: The request conflicts with current resource state or a resource limit.
      headers:
        X-Request-ID: {$ref: "#/components/headers/RequestID"}
        Cache-Control: {$ref: "#/components/headers/NoStore"}
      content:
        application/json:
          schema: {$ref: "#/components/schemas/Error"}
    PayloadTooLarge:
      description: The request body exceeds the configured server limit.
      headers:
        X-Request-ID: {$ref: "#/components/headers/RequestID"}
        Cache-Control: {$ref: "#/components/headers/NoStore"}
      content:
        application/json:
          schema: {$ref: "#/components/schemas/Error"}
    UnsupportedMediaType:
      description: The request body is not a supported application/json representation.
      headers:
        X-Request-ID: {$ref: "#/components/headers/RequestID"}
        Cache-Control: {$ref: "#/components/headers/NoStore"}
      content:
        application/json:
          schema: {$ref: "#/components/schemas/Error"}
    ValidationError:
      description: The request is syntactically valid but violates resource validation rules.
      headers:
        X-Request-ID: {$ref: "#/components/headers/RequestID"}
        Cache-Control: {$ref: "#/components/headers/NoStore"}
      content:
        application/json:
          schema: {$ref: "#/components/schemas/Error"}
...
    ServiceUnavailable:
      description: Work was cancelled, timed out, or a required dependency is temporarily unavailable.
      headers:
        X-Request-ID: {$ref: "#/components/headers/RequestID"}
        Cache-Control: {$ref: "#/components/headers/NoStore"}
      content:
        application/json:
          schema: {$ref: "#/components/schemas/Error"}
    InternalError:
      description: An unexpected internal failure occurred without exposing implementation details.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestID"
        Cache-Control:
          $ref: "#/components/headers/NoStore"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            internal:
              value:
                error:
                  code: internal_error
                  message: internal server error
                  request_id: 550e8400-e29b-41d4-a716-446655440000
                  retryable: false
  schemas:
    ApplicationPortInput:
      description: Container port and optional Swarm publication settings.
      example: {container_port: 8080, published_port: 0, protocol: tcp, publish_mode: ingress}
      type: object
      additionalProperties: false
      required: [container_port]
      properties:
        container_port: {type: integer, minimum: 1, maximum: 65535, example: 8080}
        published_port: {type: integer, minimum: 0, maximum: 65535, default: 0}
        protocol: {type: string, enum: [tcp, udp], default: tcp}
        publish_mode: {type: string, enum: [ingress, host], default: ingress}
    ApplicationVolumeInput:
      description: >-
        Validated bind mount accepted as private desired state. Source and target
        must be canonical absolute paths without redundant separators, trailing
        separators, dot segments, or parent traversal; targets must be unique.
      example: {source: /srv/moduleos/data, target: /data, read_only: true}
      type: object
      additionalProperties: false
      required: [source, target]
      properties:
        source: {type: string, minLength: 1, pattern: "^/"}
        target: {type: string, minLength: 1, pattern: "^/"}
        read_only: {type: boolean, default: false}
    CreateApplicationRequest:
      description: Complete desired state for a new image-based application.
      type: object
      additionalProperties: false
      required: [name, image]
      properties:
        name: {type: string, minLength: 1, maxLength: 63, pattern: "^[a-z0-9](?:[a-z0-9-]{0,61}[a-z0-9])?$", example: api}
        project_slug: {type: string, minLength: 1, maxLength: 48, pattern: "^[a-z0-9]+(?:-[a-z0-9]+)*$", default: root, example: root}
        image: {type: string, minLength: 1, maxLength: 512, example: "nginx:1.27"}
        replicas: {type: integer, minimum: 1, maximum: 20, default: 1, example: 2}
        env_vars:
          type: object
          maxProperties: 200
          additionalProperties: {type: string}
          example: {MODE: production}
        ports:
          type: array
          maxItems: 20
          items: {$ref: "#/components/schemas/ApplicationPortInput"}
          example: [{container_port: 8080}]
        volumes:
          type: array
          maxItems: 20
          items: {$ref: "#/components/schemas/ApplicationVolumeInput"}
          example: [{source: /srv/moduleos/data, target: /data, read_only: true}]
        expose: {type: boolean, default: false, example: false}
        ingress_container_port: {type: integer, minimum: 0, maximum: 65535, default: 0, example: 0}
...
    RedactedValue:
      description: Marker proving that a configured value exists without exposing it.
      example: {redacted: true}
      type: object
      additionalProperties: false
      required: [redacted]
      properties:
        redacted: {type: boolean, const: true}
    ApplicationPort:
      description: Normalized public port configuration.
      example: {container_port: 8080, published_port: 0, protocol: tcp, publish_mode: ingress}
      type: object
      additionalProperties: false
      required: [container_port, published_port, protocol, publish_mode]
      properties:
        container_port: {type: integer, minimum: 1, maximum: 65535}
        published_port: {type: integer, minimum: 0, maximum: 65535}
        protocol: {type: string, enum: [tcp, udp]}
        publish_mode: {type: string, enum: [ingress, host]}
    ApplicationVolume:
      description: Public mount descriptor that never exposes its host source path.
      example: {target: /data, read_only: true, source_redacted: true}
      type: object
      additionalProperties: false
      required: [target, read_only, source_redacted]
      properties:
        target: {type: string, minLength: 1, pattern: "^/"}
        read_only: {type: boolean}
        source_redacted: {type: boolean, const: true}
    Application:
      description: Public application state with environment values and host paths redacted.
      type: object
      additionalProperties: false
      required: [id, project_id, name, source_type, image, status, replicas, desired_run_state, desired_generation, observed_generation, observed_state, reconcile_retryable, env_vars, ports, volumes, expose, ingress_container_port, created_at, updated_at]
      properties:
        id: {type: string, format: uuid, example: 20000000-0000-4000-8000-000000000001}
        project_id: {type: string, format: uuid, example: 00000000-0000-0000-0000-000000000001}
        name: {type: string, minLength: 1, maxLength: 63, pattern: "^[a-z0-9](?:[a-z0-9-]{0,61}[a-z0-9])?$", example: api}
        source_type: {type: string, const: image, example: image}
        image: {type: string, minLength: 1, example: "nginx:1.27"}
        observed_image: {type: string, minLength: 1, example: "nginx@sha256:aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"}
        status: {$ref: "#/components/schemas/ApplicationStatus"}
        replicas: {type: integer, minimum: 0, example: 2}
        desired_run_state: {$ref: "#/components/schemas/DesiredRunState"}
        desired_generation: {type: integer, format: int64, minimum: 1, example: 3}
        observed_generation: {type: integer, format: int64, minimum: 0, example: 2}
        observed_state: {$ref: "#/components/schemas/ObservedState"}
        reconcile_error_code: {type: string, pattern: "^[a-z][a-z0-9_]{0,63}$", example: image_pull_failed}
        reconcile_retryable: {type: boolean, example: false}
        last_transition_at: {type: string, format: date-time, example: "2026-10-09T09:00:00Z"}
        last_reconciled_at: {type: string, format: date-time, example: "2026-10-09T09:00:05Z"}
        deletion_requested_at: {type: string, format: date-time, example: "2026-10-09T10:00:00Z"}
        env_vars:
          type: object
          maxProperties: 200
          additionalProperties: {$ref: "#/components/schemas/RedactedValue"}
          example: {SECRET: {redacted: true}}
        ports:
          type: array
          maxItems: 20
          items: {$ref: "#/components/schemas/ApplicationPort"}
          example: [{container_port: 8080, published_port: 0, protocol: tcp, publish_mode: ingress}]
        volumes:
          type: array
          maxItems: 20
          items: {$ref: "#/components/schemas/ApplicationVolume"}
          example: [{target: /data, read_only: true, source_redacted: true}]
        expose: {type: boolean, example: false}
        ingress_container_port: {type: integer, minimum: 0, maximum: 65535, example: 0}
        created_at: {type: string, format: date-time, example: "2026-10-09T09:00:00Z"}
        updated_at: {type: string, format: date-time, example: "2026-10-09T09:00:05Z"}
    ApplicationCollection:
      description: Bounded application list with explicit empty-array semantics.
      example: {applications: [], total: 0}
      type: object
      additionalProperties: false
      required: [applications, total]
      properties:
        applications:
          type: array
          maxItems: 10000
          items: {$ref: "#/components/schemas/Application"}
          example: []
        total: {type: integer, minimum: 0, maximum: 10000}
...
    ObservedState:
      description: Last persisted reconciliation observation for a resource.
      type: string
      enum: [pending, ready, reconciling, running, stopped, degraded, failed, deleting]
      example: ready
...
    ApplicationStatus:
      description: High-level application lifecycle status.
      type: string
      enum: [created, building, running, stopped, failed, updating]
      example: running
    DesiredRunState:
      description: Requested running or stopped application state.
      type: string
      enum: [running, stopped]
      example: running
...
    ErrorBody:
      description: Stable machine-readable details for one failed HTTP request.
      type: object
      additionalProperties: false
      required: [code, message, request_id, retryable]
      properties:
        code: {type: string, minLength: 1}
        message: {type: string, minLength: 1}
        request_id: {type: string, format: uuid}
        retryable: {type: boolean}
        fields:
          type: object
          additionalProperties: {type: string}
          example: {name: is required}
    Error:
      description: Stable envelope used by every non-stream API error.
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error: {$ref: "#/components/schemas/ErrorBody"}
      example:
        error:
          code: internal_error
          message: internal server error
          request_id: 550e8400-e29b-41d4-a716-446655440000
          retryable: false